修改配置文件的逻辑。
